Make the builds weekly
It is hard to figure out when to rebuild as there are a good set of
moving parts:
1. github.com/caddy-dns/cloudflare which is not versioned sadly
2. github.com/lucaslorentz/caddy-docker-proxy is versioned but we want
   to catch 2 versions
3. the builder image
4. the caddly alpine version

Trying to catch a change in any of the above is not trivial. Also,
generating images with no changes daily is not ideal either. We drop
that to weekly for now as a compromise.

Signed-off-by: Andrei Gherzan <andrei@gherzan.com>